# docs/openapi.contracts.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/races/{id}": {
      "get": {
        "operationId": "RacingDetailsController_race",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get race and conditions",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "RacingDetailsController_updateRace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRaceDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update race before registration closes",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/races/{id}/registrations": {
      "get": {
        "operationId": "RacingDetailsController_registrations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List race registrations",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      },
      "post": {
        "operationId": "RacingController_register",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register eligible horse for race",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/races/{id}/registrations/{registrationId}": {
      "get": {
        "operationId": "RacingDetailsController_registration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registrationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get race registration and approvals",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      },
      "patch": {
        "operationId": "RacingController_updateRegistration",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registrationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject race registration",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/races/{id}/results": {
      "get": {
        "operationId": "RacingDetailsController_results",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get race results",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/horses/{horseId}/race-results": {
      "get": {
        "operationId": "RacingDetailsController_horseResults",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get horse race history",
        "tags": [
          "racing"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/reports/dashboard": {
      "get": {
        "operationId": "ReportsController_dashboard",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role-scoped dashboard summary",
        "tags": [
          "reports"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/reports/horses/{horseId}/progress": {
      "get": {
        "operationId": "ReportsController_progress",
        "parameters": [
          {
            "name": "horseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get horse training progress report",
        "tags": [
          "reports"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
    "/api/v1/reports/club/training": {
      "get": {
        "operationId": "ReportsController_clubTraining",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get club training activity report",
        "tags": [
          "reports"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/reports/club/medical": {
      "get": {
        "operationId": "ReportsController_clubMedical",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get club medical activity report",
        "tags": [
          "reports"
        ],
        "x-implementation-status": "contract-only"
      }
    },
    "/api/v1/reports/club/finance": {
      "get": {
        "operationId": "ReportsController_clubFinance",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "501": {
            "description": "Contract only"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get later-phase manual finance summary",
        "tags": [
          "reports"
        ],
        "x-implementation-status": "contract-only"
      }
    },
...
  },
  "info": {
    "title": "Racehorse API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateRaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "conditions": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {
          "horseId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "horseId"
        ]
      },
      "UpdateRegistrationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OWNER_APPROVED",
              "MANAGER_CONFIRMED",
              "REJECTED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}